% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkar.R
\name{read_1ka}
\alias{read_1ka}
\title{Read the .sps and .txt files from 1ka}
\usage{
read_1ka(
  file_name_sps,
  file_name_txt = NULL,
  user_na = TRUE,
  as.tibble = FALSE
)
}
\arguments{
\item{file_name_sps}{File name of a .sps file}

\item{file_name_txt}{File name of a .txt file. If empty, it changes the file_name_sps file name accordingly (adds '_podatki' and changes the extension from 'sps' to 'txt').}

\item{user_na}{If the missing values should be replaced with NA (default) or left with their original values.}

\item{as.tibble}{If the returning dataframe should be returned as tibble (provides additional variable information).}
}
\value{
The data frame of read data.
}
\description{
Read the .sps and .txt files from 1ka
}
\examples{
data <- read_1ka('data.sps')
data <- read_1ka('data.sps', 'data_podatki.txt', user_na=FALSE)
data <- read_1ka('data.sps', as.tibble=TRUE)
}
